FSN-421: Remove actions/cache due to decommission
GitHub action `actions/cache` deprecated versions earlier than 3, and
decommissioned them starting February 1st, causing out pipeline to fail.
See https://github.blog/changelog/2024-12-05-notice-of-upcoming-releases-and-breaking-changes-for-github-actions/#actions-cache-v1-v2-and-actions-toolkit-cache-package-closing-down.
This action is used to cache golang dependencies, however the
actions/setup-go offers caching capabilities natively, so we can rely on
that instead to simplify the workflow. A side effect of a working
chaching is that tests start executing faster, so we need to increase
the deadline of integration test execution, to wait for the docker
service dependencies to be up and running.